requirements.txt
python-dotenv
aiogram
pymongo
web3
bs4
cffi